refactor: station 관련 package 구조 변경
