+++ src/main/java/subway/station/domain/StationRepository.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
